Core/Misc: fix some warnings. As a result removed one redundant hook from FormulaScript.
--HG--
branch : trunk

+++ src/server/game/Reputation/ReputationMgr.cpp
@@ -267,7 +267,7 @@ bool ReputationMgr::SetReputation(FactionEntry const* factionEntry, int32 standi
                 if (m_player->GetReputationRank(repTemplate->faction[i]) <= ReputationRank(repTemplate->faction_rank[i]))
                 {
                     // bonuses are already given, so just modify standing by rate
-                    int32 spilloverRep = standing * repTemplate->faction_rate[i];
+                    int32 spilloverRep = int32(standing * repTemplate->faction_rate[i]);
                     SetOneFactionReputation(sFactionStore.LookupEntry(repTemplate->faction[i]), spilloverRep, incremental);
                 }
             }
@@ -275,7 +275,7 @@ bool ReputationMgr::SetReputation(FactionEntry const* factionEntry, int32 standi
     }
     else
     {
-        float spillOverRepOut = standing;
+        float spillOverRepOut = float(standing);
         // check for sub-factions that receive spillover
         SimpleFactionsList const* flist = GetFactionTeamList(factionEntry->ID);
         // if has no sub-factions, check for factions with same parent
